Color index bug fix.

+++ Assets/Scripts/Brick.cs
@@ -24,13 +24,15 @@ public class Brick : MonoBehaviour {
         transform.localScale = localScale;
     }
 
-    public void CreateMesh(LdColorTable colorTable, BrickMesh brickMesh, bool invertNext, byte parentBrickColor, int maxStudCnt = 6)
+    public void CreateMesh(LdColorTable colorTable, BrickMesh brickMesh, bool invertNext, 
+        short parentBrickColor, int maxStudCnt = 6)
     {
         List<Vector3> vts = new List<Vector3>();
         List<int> tris = new List<int>();
         List<Color32> colors = new List<Color32>();
 
-        brickMesh.GetMeshInfo(colorTable, invertNext, parentBrickColor, ref vts, ref tris, ref colors, maxStudCnt);
+        brickMesh.GetMeshInfo(colorTable, invertNext, parentBrickColor, 
+            ref vts, ref tris, ref colors, maxStudCnt);
 
         Mesh mesh = new Mesh();
